+++ app/domain/sim/battle/rules/undead.rb
@@ -6,7 +6,7 @@ module Sim
 
         def general_alive?(combatants)
           Array(combatants).any? do |entry|
-            next false if entry[:current_health].to_i <= 0
+            next false if entry[:current_health].to_f <= 0
             next true if entry[:is_general]
 
             Array(entry[:attached_heroes]).any? { |hero| hero[:general] }
